move test util classes into testutil package
