+++ src/test/java/address/TestApp.java
@@ -8,7 +8,7 @@ import address.sync.cloud.CloudManipulator;
 import address.sync.cloud.model.CloudAddressBook;
 import address.util.Config;
 import address.util.GuiSettings;
-import address.util.TestUtil;
+import address.testutil.TestUtil;
 import javafx.stage.Screen;
 import javafx.stage.Stage;
