+++ src/test/java/address/model/EditPersonCommandTest.java
@@ -6,7 +6,7 @@ import address.model.datatypes.person.Person;
 import address.model.datatypes.person.ReadOnlyPerson;
 import address.model.datatypes.person.ViewablePerson;
 import address.util.Config;
-import address.util.TestUtil;
+import address.testutil.TestUtil;
 import com.google.common.eventbus.EventBus;
 import com.google.common.eventbus.Subscribe;
 import org.junit.*;
