+++ src/test/java/address/storage/StorageManagerTest.java
@@ -7,8 +7,8 @@ import address.model.UserPrefs;
 import address.model.datatypes.AddressBook;
 import address.util.Config;
 import address.util.FileUtil;
-import address.util.SerializableTestClass;
-import address.util.TestUtil;
+import address.testutil.SerializableTestClass;
+import address.testutil.TestUtil;
 import org.junit.Before;
 import org.junit.Test;
 import org.junit.runner.RunWith;
