+++ src/test/java/address/util/FilteredListTest.java
@@ -1,6 +1,7 @@
 package address.util;
 
 import address.model.datatypes.person.Person;
+import address.testutil.TestUtil;
 import address.util.collections.FilteredList;
 import javafx.collections.FXCollections;
 import javafx.collections.ListChangeListener;
