+++ src/test/java/address/util/XmlUtilTest.java
@@ -3,6 +3,8 @@ package address.util;
 import address.exceptions.DataConversionException;
 import address.model.datatypes.AddressBook;
 import address.storage.StorageAddressBook;
+import address.testutil.AddressBookBuilder;
+import address.testutil.TestUtil;
 import org.junit.Rule;
 import org.junit.Test;
 import org.junit.rules.ExpectedException;
