+++ src/test/java/guitests/GuiTestBase.java
@@ -3,10 +3,9 @@ package guitests;
 import address.TestApp;
 import address.events.EventManager;
 import address.model.datatypes.AddressBook;
-import address.model.datatypes.ReadOnlyAddressBook;
 import address.sync.cloud.model.CloudAddressBook;
 import address.testutil.TypicalTestData;
-import address.util.TestUtil;
+import address.testutil.TestUtil;
 import guitests.guihandles.MainGuiHandle;
 import guitests.guihandles.MainMenuHandle;
 import guitests.guihandles.PersonListPanelHandle;
